[371] Added the parameter `dialogKey` to `UpgraderAlert` that is used by the alert dialog.
